Moves message.ts into the chat folder
With @koran023
